+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
 - Rectangles now honor the theme's shadow color
+- Renamed `Banner.markup` to `Banner.markdown`
+- Removed the "multiline" style from Banners
+- Removed `Button.initially_disabled_for`
+- Added a `text_color` parameter to `Theme.from_colors` and `Theme.pair_from_colors`
+- `rio run` now checks that the installed version of Rio is up-to-date
 
 ## 0.7